requirements.txt
pysqlite3-binary
langchain
python-dotenv
pypdf
langchain_community
PyPDF2
langchain-groq
python-dotenv
langchain_openai
faiss-cpu
langchainhub
langchain_chroma
langchain_ollama
bs4
lark
lark-parser
streamlit-authenticator==0.1.5
streamlit==1.28.1
chromadb==0.4.17
pymongo